native: LP64 tail — .cpp unsigned long->u32, JUTConsole -sizeof int cast
- JAIGFrameSequence/JKRExpHeap/MapDraw/MarioCap .cpp: unsigned long->u32 in the
  signatures whose headers already moved to u32 (vtable/decl match on LP64).
- JUTConsole.hpp: TLinkList<JUTConsole, -sizeof(JKRDisposer)> -- size_t is 64-bit
  on host so -sizeof overflowed the int offset param; cast -(int)sizeof.

(Build also passes -Wno-narrowing to match the Metrowerks toolchain, which did
not enforce C++11 narrowing-in-aggregate-init strictness; that is a compiler-flag
match, not a source change.)

+++ src/Map/MapDraw.cpp
@@ -101,4 +101,4 @@ TMapDrawWall::TMapDrawWall(const char* name)
 {
 }
 
-void TMap::draw(unsigned long, JDrama::TGraphics*) const { }
+void TMap::draw(u32, JDrama::TGraphics*) const { }
